Introduce CLI module.
